debug_ui: Move common functions and duplicated code to common.rs
