+++ core/src/debug_ui/display_object.rs
@@ -26,6 +26,8 @@ use std::borrow::Cow;
 use std::ops::RangeInclusive;
 use swf::{Color, ColorTransform, Fixed8, Rectangle, Twips};
 
+use super::common::show_text_format;
+
 const DEFAULT_DEBUG_COLORS: [[f32; 3]; 10] = [
     [0.00, 0.39, 0.00], // "darkgreen" / #006400
     [0.00, 0.00, 0.55], // "darkblue" / #00008b
@@ -1489,51 +1491,6 @@ fn show_text_format_hover(ui: &mut Ui, tf: &TextFormat) {
         ui.style_mut().interaction.selectable_labels = true;
         show_text_format(ui, tf, false);
     });
-}
-
-pub fn show_text_format(ui: &mut Ui, tf: &TextFormat, skip_none: bool) {
-    Grid::new(ui.id().with("text_format_table"))
-        .num_columns(2)
-        .striped(true)
-        .show(ui, |ui| {
-            for (key, value) in [
-                ("Font Face", tf.font.as_ref().map(|v| v.to_string())),
-                ("Font Size", tf.size.map(|v| v.to_string())),
-                ("Color", tf.color.map(|v| format!("{v:?}"))),
-                ("Align", tf.align.map(|v| format!("{v:?}"))),
-                ("Bold?", tf.bold.map(|v| v.to_string())),
-                ("Italic?", tf.italic.map(|v| v.to_string())),
-                ("Underline?", tf.underline.map(|v| v.to_string())),
-                ("Left Margin", tf.left_margin.map(|v| v.to_string())),
-                ("Right Margin", tf.right_margin.map(|v| v.to_string())),
-                ("Indent", tf.indent.map(|v| v.to_string())),
-                ("Block Indent", tf.block_indent.map(|v| v.to_string())),
-                ("Kerning?", tf.kerning.map(|v| v.to_string())),
-                ("Leading", tf.leading.map(|v| v.to_string())),
-                ("Letter Spacing", tf.letter_spacing.map(|v| v.to_string())),
-                ("Tab Stops", tf.tab_stops.as_ref().map(|v| format!("{v:?}"))),
-                ("Bullet?", tf.bullet.map(|v| v.to_string())),
-                ("URL", tf.url.as_ref().map(|v| v.to_string())),
-                ("Target", tf.target.as_ref().map(|v| v.to_string())),
-                ("Display", tf.display.map(|v| format!("{v:?}"))),
-            ] {
-                if skip_none && value.is_none() {
-                    continue;
-                }
-
-                ui.label(key);
-                if let Some(value) = value {
-                    if !value.is_empty() {
-                        ui.label(value);
-                    } else {
-                        ui.weak("Empty");
-                    }
-                } else {
-                    ui.weak("None");
-                }
-                ui.end_row();
-            }
-        });
 }
 
 pub fn open_display_object_button<'gc>(
